dom/tests/mochitest/bugs/test_bug327891.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=327891
-->
<head>
  <title>Test for Bug 327891</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=327891">Mozilla Bug 327891</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

var pass = false;

try {
  new HTMLDocument();
} catch(e) {
  pass = true;
}

ok(pass, "new HTMLDocument() should have thrown");

</script>
</pre>
</body>
</html>
